Paulo: Descontos Finalizados

+++ ex4-pizzaria-base/src/main/java/com/bcopstein/ex4_lancheriaddd_v1/Adaptadores/Dados/PedidoRepositoryJDBC.java
@@ -98,12 +98,12 @@ public class PedidoRepositoryJDBC implements PedidoRepository{
         return count != null ? count : 0;
     }
     // Retorna o valor gasto nos últimos 30 dias
-    public float valorGastoUltimos30Dias(String clienteCpf){
+    public Double valorGastoUltimos30Dias(String clienteCpf){
         String sql = "SELECT COALESCE(SUM(valorCobrado), 0) FROM pedidos " +
                      "WHERE cliente_cpf = ? AND dataHoraPagamento >= CURRENT_DATE - INTERVAL '30' DAY";
 
         Number soma = jdbcTemplate.queryForObject(sql, Number.class, clienteCpf);
-        return soma == null ? 0f : soma.floatValue();
+        return soma == null ? 0f : soma.doubleValue();
     }
 
     @Override
